condition wordpress config on wp-config.php existence + fix sendmail issue

+++ srcs/requirements/wordpress/tools/wp_ini.sh
@@ -1,18 +1,17 @@
 #!/bin/bash
-set -a
-MYSQL_PASSWORD=$(cat "$MYSQL_PASSWORD_FILE")
-source "$WORDPRESS_CREDENTIALS_FILE"
-set +a
+if [ ! -f wp-config.php ]; then
+    set -a
+    MYSQL_PASSWORD=$(cat "$MYSQL_PASSWORD_FILE")
+    source "$WORDPRESS_CREDENTIALS_FILE"
+    set +a
 
-# TODO: remove user listing
-# TODO: avoid Error displays for user creation
-# FIXME: wordpress-1  | sh: 1: /usr/sbin/sendmail: not found on first wp init
-wp-cli config create --allow-root --dbname="$MYSQL_DATABASE" --dbuser="$MYSQL_USER" --dbpass="$MYSQL_PASSWORD" --dbhost="$WORDPRESS_DB_HOST"
-wp-cli core install --allow-root --url="$FQDN" --title="$WP_BLOG_TITLE" --admin_user="$WP_ADMIN" --admin_password="$WP_ADMIN_PASSWORD" --admin_email="$WP_ADMIN_MAIL"
-wp-cli user create --allow-root "$WP_USER" "$WP_USER_MAIL" --user_pass="$WP_USER_PASSWORD" --role=author
+    wp-cli config create --allow-root --dbname="$MYSQL_DATABASE" --dbuser="$MYSQL_USER" --dbpass="$MYSQL_PASSWORD" --dbhost="$WORDPRESS_DB_HOST"
+    wp-cli core install --allow-root --url="$FQDN" --title="$WP_BLOG_TITLE" --admin_user="$WP_ADMIN" --admin_password="$WP_ADMIN_PASSWORD" --admin_email="$WP_ADMIN_MAIL"
+    wp-cli user create --allow-root "$WP_USER" "$WP_USER_MAIL" --user_pass="$WP_USER_PASSWORD" --role=author
 
-# Permalink Setting
-wp-cli rewrite structure '/%year%/%monthnum%/%day%/%postname%/' --allow-root
+    # Permalink Setting
+    wp-cli rewrite structure '/%year%/%monthnum%/%day%/%postname%/' --allow-root
+fi
 
 # tail -f
 # Start PHP-FPM in the foreground
